+++ mcpython/common/block/Carpet.py
@@ -37,10 +37,7 @@ class AbstractCarpet(mcpython.common.block.AbstractBlock.AbstractBlock, ABC):
         instance: mcpython.common.block.AbstractBlock.AbstractBlock = dim.get_block(
             (x, y - 1, z)
         )
-        if instance is None or (
-            type(instance) != str
-            and not instance.face_solid[0]
-        ):
+        if instance is None or (type(instance) != str and not instance.face_solid[0]):
             dim.get_chunk_for_position((x, y, z)).remove_block(
                 (x, y, z), block_update=False
             )