+++ mcpython/common/network/packages/ServerChangePackage.py
@@ -1,5 +1,5 @@
 """
-mcpython - a minecraft clone written in python licenced under the MIT-licence
+mcpython - a minecraft clone written in python licenced under the MIT-licence 
 (https://github.com/mcpython4-coding/core)
 
 Contributors: uuk, xkcdjerry (inactive)
@@ -22,6 +22,7 @@ class ServerChangePackage(AbstractPackage):
     """
     Package server -> client sending a request to change the server
     """
+
     PACKAGE_NAME = "minecraft:server_change"
 
     def __init__(self):
@@ -43,14 +44,18 @@ class ServerChangePackage(AbstractPackage):
         self.new_server_port = buffer.read_int()
 
     def handle_inner(self):
-        logger.println("[NETWORK][INFO] Preparing for server change, please stand by for new connection...")
+        logger.println(
+            "[NETWORK][INFO] Preparing for server change, please stand by for new connection..."
+        )
         shared.tick_handler.schedule_once(self.reconnect)
         shared.NETWORK_MANAGER.disconnect()
 
     def reconnect(self):
         pair = self.new_server_ip, self.new_server_port
         if not connectClient2Server(pair[0], int(pair[1])):
-            logger.println("[NETWORK][FATAL] server change FAILED. See above for reason")
+            logger.println(
+                "[NETWORK][FATAL] server change FAILED. See above for reason"
+            )
             shared.state_handler.change_state("minecraft:start_menu")
             return
 
@@ -65,4 +70,3 @@ class ServerChangePackage(AbstractPackage):
         )
 
         shared.state_handler.change_state("minecraft:server_connecting")
-