+++ mcpython/common/world/World.py
@@ -18,12 +18,12 @@ import mcpython.common.config
 import mcpython.common.data.DataPacks
 import mcpython.common.entity.PlayerEntity
 import mcpython.common.state.GameViewStatePart
-import mcpython.engine.world.AbstractInterface
 import mcpython.common.world.Chunk
 import mcpython.common.world.Dimension
 import mcpython.common.world.GameRule
 import mcpython.common.world.OffProcessWorldAccess
 import mcpython.common.world.SaveFile
+import mcpython.engine.world.AbstractInterface
 import mcpython.server.worldgen.WorldGenerationHandler
 import mcpython.util.math
 import pyglet