+++ mcpython/server/command/CommandParser.py
@@ -16,8 +16,8 @@ import typing
 import mcpython.common.data.DataPacks
 import mcpython.server.command.Builder
 from mcpython import shared
-from mcpython.engine.world.AbstractInterface import IDimension
 from mcpython.engine import logger
+from mcpython.engine.world.AbstractInterface import IDimension
 
 
 class CommandExecutionEnvironment: