+++ mcpython/server/worldgen/WorldGenerationHandler.py
@@ -16,9 +16,9 @@ import typing
 
 import mcpython.common.event.Registry
 import mcpython.common.mod.ModMcpython
-import mcpython.engine.world.AbstractInterface
 import mcpython.common.world.Chunk
 import mcpython.common.world.Dimension
+import mcpython.engine.world.AbstractInterface
 import mcpython.server.worldgen.layer.ILayer
 import mcpython.server.worldgen.map.AbstractChunkInfoMap
 import mcpython.server.worldgen.mode